Add dask runner getter to credentials block and remove PC runner class
